backend/requirements.txt
langchain-community==0.2.16
langchain>=0.2.16
transformers>=4.39.0
huggingface_hub>=0.23.0
fastapi==0.95.0
uvicorn==0.22.0
requests==2.31.0
sentence-transformers==3.0.1
python-dotenv==1.0.0
openai==0.28.0
tokenizers>=0.19.1
